Remove doi_set task
Since the doi_set task returns all of the DOIs to harvest, it also puts
them in Xcoms, and its hundreds of thousands of DOIs. The recommended
practice is that Xcoms not be used for passing large amounts of data
between tasks, but just for metadata, in this case the path to the
doi-sunet.pickle file that contains the DOIs.

Fixes #149