update README instructions for getting credentials
noticed just now when looking for MAIS creds (after noticing that the related tests were skipped for me locally): these instructions were using an outdated environment name.  i fixed the shell script to use `stage` instead of `dev`, added the MAIS creds to vault (for `stage` only) so that the script would output them, and added a pointer to where the URL that's needed can be found.