Added distill step
This PR adds a distill step to the DAG which will update the publication
table for the title, pub_year and open_access.

We need to distill more metadata including:

- apc
- funders
- federally_funded

Since the `type` value is going to simply mix together the Dimensions
and OpenAlex types into a list I propose we collect those as part of the
publishing process.

Closes #216