OpenAlex query fix
OpenAlex doesn't currently support filter querying DOIs, or any values,
if they contain a comma. Until this is fixed we need to prevent querying
OpenAlex for DOIs that happen to contain a comma.

Similarly if the DOI starts with the string `doi:` that gets interpreted
as trying to do an OR query and causes API to return an error.

Fixes #277