prod docker compose: build and use the image from the local copy of the branch
instead of using the image pushed to docker hub.  this makes it easier to deploy a branch with dependency updates (whether for testing or production), since the dependencies are baked into the image, and since we don't auto-update the image from CI (since deployment happens to an on-prem VM using capistrano).