Handle invalid integer values
The publication year values can sometimes be strings instead of
integers. This commit adds an `is_int` property to JsonPathRule which
will cause it to case the matched value to an integer if found, and None
if it isn't an integer. All possible rules will be tried before
returning None. So if a sulpub value is a non-integer the dimensions
value will be tried next.

The unexpected values are logged. But we may need to add more context to
track them down if there turn out to be a lot of them.

Fixes #279