Don't query for DOIs containing pmcid:
When querying for multiple DOIs, and one of the DOIs contained the
substring `pmcid:` we encountered an OpenAlex error:

```
pyalex.api.QueryError: It looks like you're trying to do an OR query between filters and it's not supported.
You can do this: institutions.country_code:fr|en, but not this: institutions.country_code:gb|host_venue.issn:0957-1558.
```

We need to ignore these during the fillin step. We received a DOI like
this from sulpub prod:

sulpub_id: 895752
doi: 10.1093/noajnl/vdad070.013 pmcid: pmc10402389