Join Funders table
The school and department queries report on data from the Funders table
without joining on it, which results in a cartesian product that fills
up the database sort space. This commit adds the join, and adjusts the
tests to sort by the school and department as appropriate to ensure that
the rows can be reliably tested (they are not dependent on the implicit sort
order of the query results).

Fixes #423 🤞

This PR also includes several changes to publication_utils.py to improve
the reliability of working with the extracted JSON, as well as a
performance improvement to the distiller, which now caches JSON Paths
that have been parsed.