Build a static website
This commit adds a publish_website DAG that will generate a static
website using the https://github.com/sul-dlss-labs/rialto-site
repository and publish it to the website subdirectory of $DATA_DIR

The built static website currently is made available using a Airflow plugin
that serves it from the `data/website` directory. You can find a link to
the site in the Airflow menu on the far right: the link `Site`.