Normalize spaces in DOIs and prevent lookup
We had a situation where the DOI "doi:10.1016/ j.xkme.2022.100435"
was being looked up. Our normalize_doi() function was stripping leading
and trailing whitespace but not whitespace inside the string. This
change adds the stripping of whitespace, and also ensures that we don't
try to look them up in the Crossref API which throws a HTTP 400 error
that stops the workflow.