initial implementation of database tables for reporting and visualization
the intent is for this to gradually replace CSV files stored in google drive as a source of data for driving things like tableau visualizations and other reporting.  the data is distilled and denormalized, for easier querying, from the more complex (and differently duplicative) data obtained directly from harvesting.  as a start, we'll gradually add tables with the same data as the CSVs we're generating, while continuing to generate the CSVs in parallel.  we intend to remove the CSV generation, and the tables will likely end up drifting from the CSV structures they're replacing, since postgres may allow richer and more flexible querying.

the report data will be stored in a new schema with a stable name, where the data in the tables is repopulated on each run (unlike harvesting, where we create a new DB every harvest run).