Add constraints for WOS and Dimensions ID
This commit illustrates how we can reduce duplication by adding unique
constraints for platform specific identifiers like the Web of Science ID
and Dimensions ID.

First it adds unique indexes:

- doi_wos_idx: (DOI, wos ->> 'UID')
- doi_dim_idx: (DOI, wos ->> 'id')

The DOI value is coalesced into an empty string, so that the index
continues to work when a DOI is null.

Then when we are doing the "upsert" during harvesting we reference the
relevant index, to catch unique constraint violations.

Unfortunately only one index can be used as a constraint during an
upsert. Also, it only appears to be possible to reference an unique
constraint by name, so the index name is looked up manually with
get_index().