Generate publications by school
This commit adds steps for generating publications_by_school and
publications_by_department tables to the reports database for Tableau.

I had initially hoped to get some code reuse out of the very similar
queries that group publications by author, college and department. But
it wasn't immediately obvious how to do that, and I thought it might
introduce more complexity that made the queries harder to understand. We
at least do get reuse of the dataset fixture though.

I tested with a harvest limit of 1K in my development environment and
compared the results with the equivalent openaccess CSV output files.
They were the same except we aren't limiting to pub_year >= 2018.

Closes #521 Closes #522