Move types generation into distill
Previously we were doing some complicated logic in SQL and Python to
generate the list of publication types for our reporting. This commit
moves it into the distill step instead, which results in a new column
"types" on the "publication" table, which can be null or a list of
string values.

Closes #585