rialto_airflow.harvest.wos: _wos_api retry logic centralized and tweaked
centralize the Retry configuration for two _wos_api GET requests that instantiated Retry objects with the same params.

update the Retry params to use a longer backoff_factor, to add a backoff_jitter, and to retry twice as many times.  hopefully the longer backoff iterations will only be necessary occasionally (when the API rate limits us for a while).  a longer backoff should allow the rate limiting block to subside, and should hopefully allow further requests to proceed without having to retry at all for a while.

it doesn't seem like it will work to catch a particular 429 exception, and alert/proceed, because presumably the next request will get a 429 also (since the issue is too many requests too quickly, not the particular requests themselves).