Handle empty string in arxiv doi extraction
It looks like empty string DOI (or DOIs solely comprised of spaces) are
being passed to normalize_doi() which then causes
normalize_arxiv_id_to_doi() to return None, since "" evaluates to False.
This then causes the doi_candidate_extract() to fail since it isn't
expecting a DOI of None.

The solution here is to have noramlize_arxiv_id_to_doi() to always
expect a str, it should blow up if passed a None. A check early in
normalize_doi was added to make sure empty string, or strings of just
spaces are normalized to None.

Fixes #696